Simplify h norm calculation.
torch.linalg.vector_norm works on floating-point types, not int or long,
so additionally ensure that the default Action and FermionMatrix µ and h
are floats.

+++ python/tdg/tdg/action.py
@@ -7,7 +7,7 @@ from tdg.fermionMatrix import FermionMatrix
 
 class Action:
 
-    def __init__(self, spacetime, potential, beta, mu=torch.tensor(0), h=torch.tensor([0,0,0]), fermion=FermionMatrix):
+    def __init__(self, spacetime, potential, beta, mu=torch.tensor(0, dtype=torch.float), h=torch.tensor([0,0,0], dtype=torch.float), fermion=FermionMatrix):
         self.Spacetime = spacetime
         self.Potential = potential
